requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0

channels==4.3.2
daphne==4.2.1

djangorestframework-simplejwt==5.5.1
PyJWT==2.10.1

python-dotenv==1.2.1
python-decouple==3.8

asgiref==3.11.0
sqlparse==0.5.4
tzdata==2025.2
typing_extensions==4.15.0
assemblyai==0.48.1
psycopg2-binary==2.9.10
